feat: move dojo cnotext so that avoid unnecessary loading in lp
