+++ packages/client/src/app/(game)/game/[id]/page.tsx
