+++ packages/client/src/app/(game)/game/_components/StageMenuHeader.tsx
