+++ packages/client/src/app/(game)/game/page.tsx
