+++ packages/client/src/app/(game)/stage/[id]/edit/page.tsx
