+++ packages/client/src/app/providers.tsx
@@ -2,62 +2,13 @@
 
 import { mainnet, sepolia } from '@starknet-react/chains'
 import { StarknetConfig, publicProvider, argent, braavos, voyager } from '@starknet-react/core'
-import { useEffect, useState } from 'react'
 import { ArgentMobileConnector } from 'starknetkit/argentMobile'
 import { WebWalletConnector } from 'starknetkit/webwallet'
-import { dojoConfig } from '@/../dojoConfig'
 import { Toaster } from '@/components/ui/sonner'
 import { APP_DESCRIPTION, APP_NAME } from '@/constants'
-import { DojoProvider } from '@/libs/dojo/DojoContext'
-import { setup, type SetupResult } from '@/libs/dojo/setup'
 import { detectMobile } from '@/utils/devices'
 
 const Providers = ({ children }: { children: React.ReactNode }) => {
-  const [setupResult, setSetupResult] = useState<SetupResult | null>(null)
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState<Error | null>(null)
-
-  useEffect(() => {
-    const setupDojo = async () => {
-      try {
-        setIsLoading(true)
-        const setupResult = await setup(dojoConfig)
-        setSetupResult(setupResult)
-      } catch (err) {
-        console.error('Error setting up Dojo:', err)
-        setError(err instanceof Error ? err : new Error('Unknown error occurred'))
-      } finally {
-        setIsLoading(false)
-      }
-    }
-
-    setupDojo()
-  }, [])
-
-  if (isLoading) {
-    return (
-      <div className='flex h-screen items-center justify-center bg-gray-800 text-xl text-white'>
-        Loading...
-      </div>
-    )
-  }
-
-  if (error) {
-    return (
-      <div className='flex h-screen items-center justify-center bg-gray-800 text-xl text-white'>
-        Error: {error.message}
-      </div>
-    )
-  }
-
-  if (!setupResult) {
-    return (
-      <div className='flex h-screen items-center justify-center bg-gray-800 text-xl text-white'>
-        Dojo setup failed
-      </div>
-    )
-  }
-
   const chains = [mainnet, sepolia]
   const connectors = detectMobile()
     ? [
@@ -77,10 +28,8 @@ const Providers = ({ children }: { children: React.ReactNode }) => {
       explorer={voyager}
       autoConnect={true}
     >
-      <DojoProvider value={setupResult}>
-        <Toaster richColors position='top-center' />
-        {children}
-      </DojoProvider>
+      <Toaster richColors position='top-center' />
+      {children}
     </StarknetConfig>
   )
 }
